requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.33.0
django-crispy-forms==1.6.1
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
djangorestframework==3.7.3

